chore: code suggestions
