+++ crates/rome_js_analyze/tests/suppression/correctness/noDoubleEquals/noDoubleEquals.jsx
@@ -1,0 +1,7 @@
+let a = <button
+    className="SomeManyClasses"
+    onClick={(event) => console.log(event.ctrlKey == true)}
+    style="color: red"
+>
+    Some content
+</button>;
